Atualização

+++ cobranca_bb.py
@@ -9,28 +9,31 @@ def ler_cobrancabb():
 
 async def cobrancabb():
     global arquivo_cobrancabb
-    pasta = Book(arquivo_cobrancabb)
-    planilha = pasta.sheets[0]
+    if arquivo_cobrancabb == None:
+        print('Vazio')
+    else:
+        pasta = Book(arquivo_cobrancabb)
+        planilha = pasta.sheets[0]
 
-    execute_query(connection, 'USE fluxodecaixa;')
+        execute_query(connection, 'USE fluxodecaixa;')
 
-    lr = planilha.range('A2').end('down').row
+        lr = planilha.range('A2').end('down').row
 
-    soma = 0
-    for i in range(3, lr + 1):
-        data = planilha.range(f'A{i}').value
-        data1 = planilha.range(f'A{i + 1}').value
-        valor = planilha.range(f'E{i}').value
-        if isinstance(data, datetime):
-            soma += valor
-            if data != data1 or data1 == None:
-                data_cobrancabb.append(data)
-                valor_cobrancabb.append(soma)
-                soma = 0
+        soma = 0
+        for i in range(3, lr + 1):
+            data = planilha.range(f'A{i}').value
+            data1 = planilha.range(f'A{i + 1}').value
+            valor = planilha.range(f'E{i}').value
+            if isinstance(data, datetime):
+                soma += valor
+                if data != data1 or data1 == None:
+                    data_cobrancabb.append(data)
+                    valor_cobrancabb.append(soma)
+                    soma = 0
 
-    for i in range(0, len(data_cobrancabb)):
-        execute_query(connection, "INSERT INTO bbrasil (data, valor) VALUES ('{}', '{}');".format(data_cobrancabb[i], valor_cobrancabb[i]))
+        for i in range(0, len(data_cobrancabb)):
+            execute_query(connection, "INSERT INTO bbrasil (data, valor) VALUES ('{}', '{}');".format(data_cobrancabb[i], valor_cobrancabb[i]))
 
-    pasta.close()
-    os.system('taskkill /im Excel.exe')
-    await asyncio.sleep(0)
+        pasta.close()
+        os.system('taskkill /im Excel.exe')
+        await asyncio.sleep(0)
